requirements.txt
pandas==1.5.1
azure-identity==1.12.0
azure-mgmt-core==1.3.2
azure-mgmt-storage==21.0.0
azure-storage-blob==12.14.1
psycopg2-binary==2.9.5